chore(gatsby): migrate type of filter to TypeScript
