+++ packages/gatsby/src/schema/schema.js
@@ -36,7 +36,7 @@ const {
 } = require(`./extensions`)
 import { getPagination } from "./types/pagination"
 const { getSortInput, SORTABLE_ENUM } = require(`./types/sort`)
-const { getFilterInput, SEARCHABLE_ENUM } = require(`./types/filter`)
+import { getFilterInput, SEARCHABLE_ENUM } from "./types/filter"
 const { isGatsbyType, GatsbyGraphQLTypeKind } = require(`./types/type-builders`)
 const {
   isASTDocument,
