+++ packages/gatsby/src/schema/types/__tests__/sort-and-filter.js
@@ -1,7 +1,7 @@
 // NOTE: Previously `infer-graphql-input-from-fields-test.js`
 
 const { createSchemaComposer } = require(`../../schema-composer`)
-const { getFilterInput } = require(`../filter`)
+import { getFilterInput } from "../filter"
 const { getSortInput } = require(`../sort`)
 
 const {
